Convert event handling to select
The previous command with smol::or was not correct because it could
deadlock - just because the socket has pending data doesn't mean it has
a full PDU readable. The event processing would stop until the full PDU
is read.

This seems to work in practice with TLS, but with Quic it does not
because the server can't accept the connection until the client sends
something, but the client can't send because its waiting for a PDU from
the server (even though its just Quic-internal traffic that's making the
socket readable).

We also use StreamPeek to determine whether the socket is readable by checking
there's at least one byte of user-level data ready to read. Once
we see this we assume the rest of the PDU will arrive in bounded time, so we
can call Pdu::decode_async without risk of it being cancelled.